requirements.txt
# Arabic OCR Project Dependencies
# Tested and verified working configuration

# Core deep learning framework (CPU-only version for better Windows compatibility)
tensorflow-cpu>=2.16.1

# Computer vision and image processing
opencv-python>=4.5.0
Pillow>=8.3.0

# Scientific computing
numpy>=1.21.0,<2.0.0

# GUI framework
PyQt6>=6.0.0

# Utilities
matplotlib>=3.5.0

# Model compatibility
h5py>=3.1.0